requirements.txt
streamlit
langchain
langchain-groq
googlesearch-python
beautifulsoup4
requests
python-dotenv